fix: 업무분야 인덱스 페이지 인터페이스, 스타일 분리
