+++ src/components/work/work.interface.tsx
@@ -1,0 +1,10 @@
+export interface Category {
+  categoryId: `C${number}`;
+  name: string;
+  category: SubCategory[];
+}
+
+export interface SubCategory {
+  categoryId: `S${number}`;
+  name: string;
+}
