+++ src/components/work/work.styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Grid = styled.div`
+  width: 100%;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+`;
